v1.72.0 solucionado el problema de los decimales en la migration more vue productoHome

+++ database/migrations/2022_05_10_084326_create_tbl_producto_table.php
@@ -21,8 +21,10 @@ return new class extends Migration
             
             // $table->morphs('tokenable');
             $table->string('nombre');
-            $table->double('precio',50,2);
-            $table->double('newPrecio',50,2);
+            // $table->double('precio',50,2);
+            // $table->double('newPrecio',50,2);
+            $table->decimal('precio',8,2);
+            $table->decimal('newPrecio',8,2);
             $table->string('description');
             $table->string('peso');
             $table->integer('cantidad');
